Make validation matrix

This function uses the validation trials from triplet data to construct a subject-by-trial matrix of judgments on these items.

Usage

make.vmat(triplist)

Arguments

triplist

Named list whose elements each contain triplet data from one participant. Names should be participant identifiers, as returned by get_combined.

Value

Names list with two elements. majority is a matrix with one row per validation triplet, reporting the winning choice and proportion of participants who selected ite. bysbj is a matrix indicating, for each participant (rows) and validation triplet (columns), the choice the participant made.

Details

Each validation triplet is identified with a unique code of the kind generated by make.tripnames which indicates the target word and the two options, with the options ordered alphabetically.

Where a participant judged a particular validation triplet the entries of bysbj will indicate the proportion of times the participant's decision agreed with the majority vote. (In most cases this will be 0 or 1 since participants usually judge a validation triplet only once; however in some studies the same validation trials are repeated to evaluate self-consistency). Where a participant did not judge a given triplet, the entry in bysbjwill be NA.

Examples

vmat <- make.vmat(icon_triplets)
vmat$majority
#>              triplet majority pmaj
#> 1  fdfow_fnmyb_pdhos    fnmyb  1.0
#> 2  fdmob_fdfow_pnhob    fdfow  1.0
#> 3  fdmob_fdmow_fdmyb    fdmyb  0.8
#> 4  fdmob_fnmob_pdcos    fnmob  1.0
#> 5  fdmow_fdmyw_fnfyw    fdmyw  1.0
#> 6  fdmow_fnfyw_pncnb    fnfyw  1.0
#> 7  fdmow_fnmyw_pdhnb    pdhnb  0.6
#> 8  fdmyb_fdfob_pdhos    fdfob  1.0
#> 9  fdmyb_fnfyb_pdcns    fnfyb  0.6
#> 10 fdmyb_fnmyb_fnmyw    fnmyb  1.0
#> 11 fdmyw_fdfow_fnfyw    fnfyw  1.0
#> 12 fdmyw_fnfob_fnmyw    fnmyw  1.0
#> 13 fdmyw_pncns_pnhns    pnhns  1.0
#> 14 fnfyb_fdmyb_pdcnb    fdmyb  1.0
#> 15 fnmob_fdmob_fnmyb    fdmob  0.6
#> 16 fnmob_fdmow_pdcns    fdmow  1.0
#> 17 fnmyb_fdfob_pdcns    fdfob  0.8
#> 18 fnmyb_pnhnb_pnhos    pnhnb  1.0
#> 19 fnmyw_pdcnb_pnhos    pnhos  0.6
#> 20 pdcnb_fdmob_pdhnb    pdhnb  1.0
#> 21 pdcnb_pncos_pnhnb    pncos  1.0
#> 22 pdcns_fnmob_fnmow    fnmob  0.8
#> 23 pdcns_pdcnb_pdcob    pdcnb  1.0
#> 24 pdcns_pdcob_pdcos    pdcos  1.0
#> 25 pdcob_fdfob_pnhos    pnhos  0.6
#> 26 pdcob_fnfob_pnhob    pnhob  1.0
#> 27 pdcos_fnfow_pdcnb    pdcnb  1.0
#> 28 pdcos_fnmob_pncnb    pncnb  1.0
#> 29 pdhnb_fdfow_pdhos    pdhos  1.0
#> 30 pdhnb_fdmob_pdcnb    pdcnb  1.0
#> 31 pdhnb_pdcos_pnhos    pnhos  1.0
#> 32 pdhob_fdfow_fnmob    fdfow  1.0
#> 33 pdhob_fdfyb_pnhob    pnhob  1.0
#> 34 pdhob_pdcob_pnhns    pdcob  1.0
#> 35 pdhos_fnmow_pncnb    pncnb  1.0
#> 36 pncnb_fdmyb_pdcos    pdcos  1.0
#> 37 pncnb_fnmow_pdhob    pdhob  0.6
#> 38 pncob_fnmow_pdcos    pdcos  1.0
#> 39 pnhnb_fnmow_pncos    pncos  1.0
#> 40 pnhos_fnfob_pdhos    pdhos  1.0